Decouple l1 config from builder config
